Fixed negate with overflow in Intel formatter.
If the immediate value happened to be -MIN (e.g. -128 for an i8), negating that value would cause an `attempt to negate with overflow` crash in debug builds.
This fix uses wrapping_neg instead, which would wrap -in::MIN to in::MIN, which when cast to the unsigned type will give the expected result.