Add `ICRegister` for `Instruction.create()` methods
The `Register` is not an enum, just a class with integers, which causes
problems when overloading the `create()` method, so the param type had
to be included in the method name, eg. `create_reg_mem()` etc. This
commit adds `ICRegister` and `ICRegisters*` and updates the `create()`
methods to take an `ICRegister` instead of an `int` (a `Register`)
parameter so they can now all be called `create()`.