Fix asan reported leaks
`Box::leak(Box::new(vec)).as_ptr()` allocates the Vec itself on the
heap, then only gets the pointer to the data in the vec, so the vec
itself is unreachable. Instead, use `vec.leak()` directly, which doesn't
allocate the vec itself on the heap, just leaking the vec's data, which
we keep a pointer to.