Fix tests when compiling with `--target`
When `--target` is passed, (even if it's the same as the default
target), the executable will end up in `target/{target_tripple}/debug`.
This extra folder level broke the `get_unit_tests_base_dir` function,
which assumed the unit tests dir would be 5 levels above the executable.

Instead, just use the current directory: cargo runs tests from the same
directory as the crate root which defines the tests, therefore we can
use a relative path to the unit tests from the `iced-x86` directory